+++ src/main/java/kitchenpos/domain/Menu/Menu.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.Menu;
 
 import java.math.BigDecimal;
 import java.util.List;
